create_geneAdj_and_mutM.R
I created a function to read the maf file: createScoreTable this is a replacement for
the geneScore command in the previous capture. Also geneScores variable is now the scores variable.

I also created the getFuncMutations as an enclosed function to extract the non-slient mutations
from the output of the createScoreTable function above.


create_metadata_structure:

I enclosed the operations to combine two metadata tables within the combineMetadata function.
it orders the variables of one table according to the other table and then rbinds one table to the bottom of the other.