create_geneAdj_and_mutM.r
I have added a filtering step at lines36-38 in order to analyse the data
split according to sequecning machine. For some reason it is not working.
It has something to do with the uniqueSamples variable. It still retains
354 factors even though 266 have been given to it.

identify_hypermutated_samples.r

The Fisher's test does test for deviation from an expected distribution of
class membership given the marginal totals in the crossTB table. I also
have a test at line 146 based on the "mutations" table which tests to see
if there is a difference between the number of mutations identified by the
solid and illumina machines given the marginal totals.From line 150 I am
also working on a binomial test to see if solid and illumina sequecning
identified different number of mutations per gene. This may however need
to be stratified by cancer type.