identify_mutated_samples:
I hae modified the MDS plot script to color the points according to the
smpclrs variable, the legend construction is now automated to construct
according to the attributes of the smpclrs variable ie: then number of
unique classes in smpclrs and the text associated with each class
determines the point color and the legend title. I am also modifying the
plot text sizes and poisitions so they can be easily read in plot outputs.
This has involved moving the leged to the right of the plot with ample
room for a larger list of classes.

I now need to automate the generation of the plot within a function so I
can loop over the metadata classes I have in my stratification variable to
create as many MDS plots as I have metadata variables in the
stratification variable. I also need to put all of my modified code
together in to one package now.