identify hypermutated samples.R
I have changed hypermutated samples in order to produce the hypemutatedTB
which produces a crossTab of hypermutated against all other metadata and
or counts of mutations in genes. I am also trying to modify the metadata2
table to be a dataframe rather than a matrix so that I can identify the
"type" of the metadata variables so that in the hypermutatedTB I can apply
different statistical tests based on the number of metadata groups I have
and the type of the data.