added draft.package.R
I have added the draft package.R file which contains various lines of code
from:
create_metadata_structure.R
create_geneAdj_and_mutM.R

with a view to creating a package. I have modified the create metadata
part of the file to create colon and rectum objects that are dataframes
without factors. I need this for a later stage where I want to identify
the most appropriate statistical test for the variable based on the
datatype and number of groups. The next stage is to modify the colorectal
object to be a dataframe and modify the creation of objects based on the
colorectal object by reversing t()functions and swapping colnames for
rownames.