metadata Types
I have added a variable which tells me the type of the metadata
variables. This is based on the metadata2 variable. I have also created
a metadata3 variable based on metadata2 where NA is converted to "not
available" for the purposes of generating plot colors where NA values
are not identifiable as factors.