compdiss fixed and isoMDS plots added
In network_informed_clustering_function.R  I changed the compDiss
function so that none of the values are 0 by adding 1 to the denominator
in computing the dissimilarity between samples. This ensures that the
two most similar samples do not have a dissimilarity of zero. Previously
they would have had a dissimilarity of 0 because I would divide the
number of matches by the maximum number of matches between any two
samples, meaning the two most similar samples would have a distance of
0, number of matches/maximum number of matches.

draft package now has code in the last few lines for creating isoMDS
plots, which now work because none of the pairwise sample
dissimilarities are 0. The plots look very good and very useful for me.