added IndelSNVratio and IndelPercentage
I have added the SNVIndel ratio normally used to identify hypermutated
samples and Indel% among functinal mutations. These measures however do
not appear to work as well as the previous one I was using, the
percentage of silent mutations amongst all mutations.

The code for this is in a previous commit and I'm not sure where.