IndelSNVratio IndelPErcentage PanCan12 mods
I have modified the code for the indication of hypermutated samples to
use the correct measure of Indel:SNV ratio rather than indel:total
mutations ratio. The  ******PerIndiv measures (~ln306) are now basd on
the correct mutationClassPerIndiv dataframe detailing if mutations are
SNPs, DNPS, INS, or DEL. This should be all I need to work with the
PANcan12 data.

A new function has been added to create a sequencer centre
label(ln207-ln223). This is used as a surrogate for sequencer machine
because PANCan12 data does not include sequencer type information.

Previous versions of this code are safe to use with TCGA obtained .MAF
files. but from this point on, code is for PANCAN12 .maf and clinical
files.